chore(csm): prettify csm worker and move as a crate.
