+++ Cargo.toml
@@ -26,6 +26,7 @@ members = [
   "crates/common",
   "crates/config",
   "crates/consensus-logic",
+  "crates/csm-worker",
   "crates/crypto",
   "crates/db",
   "crates/db-store-sled",
@@ -146,6 +147,7 @@ strata-common = { path = "crates/common" }
 strata-config = { path = "crates/config" }
 strata-consensus-logic = { path = "crates/consensus-logic" }
 strata-crypto = { path = "crates/crypto", default-features = false }
+strata-csm-worker = { path = "crates/csm-worker" }
 strata-db = { path = "crates/db" }
 strata-db-store-rocksdb = { path = "crates/db-store-rocksdb" }
 strata-db-store-sled = { path = "crates/db-store-sled" }
