+++ crates/consensus-logic/Cargo.toml
@@ -7,14 +7,13 @@ version = "0.3.0-alpha.1"
 workspace = true
 
 [dependencies]
-strata-asm-common.workspace = true
-strata-asm-logs.workspace = true
 strata-asm-types.workspace = true
 strata-asm-worker.workspace = true
 strata-chain-worker.workspace = true
 strata-chainexec.workspace = true
 strata-chaintsn.workspace = true
 strata-common = { workspace = true, default-features = true }
+strata-csm-worker.workspace = true
 strata-db.workspace = true
 strata-eectl.workspace = true
 strata-ol-chain-types.workspace = true
@@ -31,7 +30,6 @@ anyhow.workspace = true
 bitcoin.workspace = true
 bitcoind-async-client.workspace = true
 borsh.workspace = true
-serde.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
