--- crates/consensus-logic/src/csm/mod.rs
@@ -1,13 +1,0 @@
-//! Consensus state machine
-//!
-//! This is responsible for managing the final view of the checkpointing state,
-//! tracking unrecognized state from L1, and determining the basis for which
-//! unfinalized blocks are committed.
-// TODO clean up this module so that specific items are directly exported and
-// modules don't have to be
-
-pub mod message;
-pub mod worker;
-
-// Re-export the new service types
-pub use worker::{CsmListenerService, CsmListenerState, CsmListenerStatus};
