+++ crates/consensus-logic/src/lib.rs
@@ -3,10 +3,10 @@
 pub mod asm_worker_context;
 pub mod chain_worker_context;
 pub mod checkpoint_verification;
-pub mod csm;
 pub mod exec_worker_context;
 pub mod fork_choice_manager;
 pub mod genesis;
+pub mod message;
 pub mod sync_manager;
 pub mod tip_update;
 pub mod unfinalized_tracker;
