+++ crates/csm-worker/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "strata-csm-worker"
+version = "0.1.0"
+edition = "2024"
+
+[lints]
+workspace = true
+
+[dependencies]
+strata-asm-common.workspace = true
+strata-asm-logs.workspace = true
+strata-asm-worker.workspace = true
+strata-common = { workspace = true, default-features = true }
+strata-db.workspace = true
+strata-primitives.workspace = true
+strata-service.workspace = true
+strata-state.workspace = true
+strata-status.workspace = true
+strata-storage.workspace = true
+
+anyhow.workspace = true
+serde.workspace = true
+tracing.workspace = true
