+++ crates/sync/src/worker.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 use futures::StreamExt;
 #[cfg(feature = "debug-utils")]
 use strata_common::{check_and_pause_debug_async, WorkerType};
-use strata_consensus_logic::{csm::message::ForkChoiceMessage, sync_manager::SyncManager};
+use strata_consensus_logic::{message::ForkChoiceMessage, sync_manager::SyncManager};
 use strata_ol_chain_types::{L2BlockBundle, L2Header};
 use strata_primitives::epoch::EpochCommitment;
 use strata_status::ChainSyncStatusUpdate;
